Add linter rule for invalid shorthand arguments
