+++ packages/eslint-plugin/src/configs/recommended.ts
@@ -2,6 +2,7 @@ export const recommendedConfig = {
   plugins: ['@griffel'],
   rules: {
     '@griffel/hook-naming': 'error',
+    '@griffel/no-invalid-shorthand-argument': 'error',
     '@griffel/no-shorthands': 'error',
     '@griffel/pseudo-element-naming': 'error',
   },
